Added WeatherManagement class to hold conversion methods and added Geography class to hold properties for constructing a geocode for an address. Address lat and long properties changed to float.

+++ Rolory/Models/Address.cs
@@ -25,7 +25,7 @@ namespace Rolory.Models
         public int? ZipCode { get; set; }
         [Display(Name = "Country")]
         public string CountryName { get; set; }
-        public int? Latitude { get; set; }
-        public int? Longitude { get; set; }
+        public float? Latitude { get; set; }
+        public float? Longitude { get; set; }
     }
 }
